[Refactor] BREAKING CHANGES: exceptions module is now public.
