+++ job_shop_lib/__init__.py
@@ -1,12 +1,6 @@
 """Contains the main data structures and base classes.
 """
 
-from job_shop_lib._exceptions import (
-    JobShopLibError,
-    NoSolutionFoundError,
-    ValidationError,
-    UninitializedAttributeError,
-)
 from job_shop_lib._operation import Operation
 from job_shop_lib._job_shop_instance import JobShopInstance
 from job_shop_lib._scheduled_operation import ScheduledOperation
@@ -21,8 +15,4 @@ __all__ = [
     "Schedule",
     "Solver",
     "BaseSolver",
-    "JobShopLibError",
-    "NoSolutionFoundError",
-    "ValidationError",
-    "UninitializedAttributeError",
 ]
