+++ job_shop_lib/exceptions.py
