+++ job_shop_lib/generation/_general_instance_generator.py
@@ -2,7 +2,8 @@
 
 import random
 
-from job_shop_lib import JobShopInstance, Operation, ValidationError
+from job_shop_lib import JobShopInstance, Operation
+from job_shop_lib.exceptions import ValidationError
 from job_shop_lib.generation import InstanceGenerator
