+++ job_shop_lib/visualization/__init__.py
@@ -9,7 +9,9 @@ from job_shop_lib.visualization._gantt_chart_video_and_gif_creation import (
     create_video_from_frames,
     create_gif_from_frames,
 )
-from job_shop_lib.visualization._disjunctive_graph import plot_disjunctive_graph
+from job_shop_lib.visualization._disjunctive_graph import (
+    plot_disjunctive_graph,
+)
 from job_shop_lib.visualization._agent_task_graph import (
     plot_agent_task_graph,
     three_columns_layout,
