+++ tests/test_operation.py
@@ -1,5 +1,6 @@
 import pytest
-from job_shop_lib import Operation, JobShopLibError
+from job_shop_lib import Operation
+from job_shop_lib.exceptions import UninitializedAttributeError
 
 
 def test_init_single_machine():
@@ -21,5 +22,5 @@ def test_machine_id_single():
 
 def test_machine_id_multiple():
     op = Operation([1, 2], 10)
-    with pytest.raises(JobShopLibError):
+    with pytest.raises(UninitializedAttributeError):
         _ = op.machine_id
